hushh-desktop/backend/requirements.txt
# Generated from pyproject.toml + uv.lock via scripts/sync_runtime_requirements.sh
# Contributor and CI installs must use: uv sync --frozen --group dev

aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
asgiref==3.11.1
asyncpg==0.31.0
attrs==26.1.0
authlib==1.6.11
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.90
botocore==1.42.90
cachecontrol==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==46.0.7
curl-cffi==0.15.0
defusedxml==0.7.1
deprecated==1.3.1
distro==1.9.0
docstring-parser==0.18.0
fastapi==0.136.0
firebase-admin==7.1.0
flask==3.1.3
frozendict==2.4.7
frozenlist==1.8.0
google-adk==1.28.1
google-api-core==2.29.0
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-aiplatform==1.148.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.27.0 ; platform_python_implementation != 'PyPy'
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-core==1.2.31
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
limits==5.8.0
mako==1.3.11
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.4
openai==2.32.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pandas==3.0.2
pdfminer-six==20231228
pdfplumber==0.11.0
peewee==4.0.4
phonenumbers==9.0.28
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycountry==26.2.16
pycparser==3.0 ; implementation_name != 'PyPy'
pydantic==2.13.1
pydantic-core==2.46.1
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.0.0
pyparsing==3.3.2
pypdfium2==5.7.0
pytesseract==0.3.13
python-a2a==0.5.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.26
pytz==2026.1.post1
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.50.0
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1 ; sys_platform == 'emscripten' or sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.44.0
watchdog==6.0.0
websockets==15.0.1
werkzeug==3.1.8
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
yfinance==1.3.0
zipp==3.23.1
zstandard==0.25.0
